fix: u16 -> u32

+++ src/client/map.h
@@ -313,7 +313,7 @@ private:
 
     void removeUnawareThings();
 
-    uint16_t getBlockIndex(const Position& pos) { return ((pos.y / BLOCK_SIZE) * (65536 / BLOCK_SIZE)) + (pos.x / BLOCK_SIZE); }
+    uint32_t getBlockIndex(const Position& pos) { return ((pos.y / BLOCK_SIZE) * (65536 / BLOCK_SIZE)) + (pos.x / BLOCK_SIZE); }
 
     std::vector<FloorData> m_floors;
